backend/requirements.txt
aioredis==1.3.1
redis==4.3.4
elasticsearch[async]==7.9.1
fastapi==0.85.0
orjson==3.8.0
pydantic==1.9.0
gunicorn==20.1.0
uvicorn==0.15.0
uvloop==0.17.0
elasticsearch-dsl==7.4.0
python-dotenv==0.21.0
python-logstash==0.4.8
PyJWT==2.6.0